requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
python-jose
passlib
pytest
alembic
python-dotenv